Move PageRevisionModelFiltered to services/revision/structs.rs
